Add Wan SCAIL workflow
Major improvements to video source handling, SCAIL workflow fidelity, and UI parity with ComfyUI for video-based workflows:
- Introduced VideoSourceOptions to unify video loader settings (rate, cap, skip, format, etc.) across UI and backend, ensuring strict parity with ComfyUI defaults.
- Enhanced VideoInput component and preview: all loader options now user-editable; preview supports remote files, frame seeking, and accurate cut summaries.
- Added SCAIL-specific workflow fragments (pose detection, rendering, embeds, sampler) and corresponding UI forms for full user control and parameter exposure.
- WanSCAILEmbedsWorkflow and related workflows now match source ComfyUI graphs in node types, dataflow, and defaults; registry and node wiring updated for output/index parity.
- Enforced a source fidelity contract: updated documentation and planning process to require fidelity matrices, deviation tracking, and user approval for any dropped nodes.
- Improved Civitai DTOs and entity models for robustness to partial API responses.
- Fixed UI and registry bugs; ensured all loader, sampler, and conditioning nodes use source-matching names, outputs, and defaults.
- Added comprehensive tests for SCAIL workflows and detailed plan.md documenting node mapping and fidelity review.
- Added new ComfyUI workflow and node preset JSONs for WanVideo 2.1 14B SCAIL pose control, providing a ready-to-use, fully documented example pipeline for pose-controlled video generation.

+++ BlazorWebApp/Workflows/Fragments/Core/LoadVideoFragment.cs
@@ -21,7 +21,7 @@ public class LoadVideoFragment : IFragmentBuilder
     public class Parameters
     {
         public string Video { get; set; } = "";
-        public int ForceRate { get; set; } = 16;
+        public double ForceRate { get; set; } = 16;
         public int CustomWidth { get; set; } = 480;
         public int CustomHeight { get; set; } = 832;
         public int FrameLoadCap { get; set; } = 176;
@@ -41,7 +41,7 @@ public class LoadVideoFragment : IFragmentBuilder
         BuildInternal(builder, registry, new Parameters
         {
             Video = fragment?.GetString("video", "") ?? "",
-            ForceRate = fragment?.GetInt("force_rate", 16) ?? 16,
+            ForceRate = fragment?.GetDouble("force_rate", 16) ?? 16,
             CustomWidth = fragment?.GetInt("custom_width", 480) ?? 480,
             CustomHeight = fragment?.GetInt("custom_height", 832) ?? 832,
             FrameLoadCap = fragment?.GetInt("frame_load_cap", 176) ?? 176,
